Clean up some unused stuff
